Extract determineDryRunSettings and export it
